SER-340 finished mercury frontend
