+++ front/main/app/models/image-review/index.js
@@ -1,0 +1,121 @@
+import Ember from 'ember';
+import rawRequest from 'ember-ajax/raw';
+import request from 'ember-ajax/request';
+
+const ImageReviewModel = Ember.Object.extend({
+	showSubHeader: true,
+
+	setImagesCount(status) {
+		request(M.getImageReviewServiceUrl('/monitoring', {
+			status
+		})).then((promise) => {
+			this.set('imagesToReviewCount', promise.countByStatus);
+		});
+	}
+});
+
+ImageReviewModel.reopenClass({
+
+	startSession(status) {
+		const options = {
+			status
+		};
+
+		return rawRequest(M.getImageReviewServiceUrl(`/batch`, options), {
+			method: 'POST'
+		}).then(({payload, jqXHR}) => {
+			// In case there are no more images, create empty model and show `No more images to review` message
+			if (jqXHR.status === 204) {
+				return ImageReviewModel.createEmptyModelWithPermission(status);
+			} else if (jqXHR.status == 404) {
+				return ImageReviewModel.createEmptyModelWithPermission(status);
+			} else {
+				return ImageReviewModel.getImagesAndPermission(payload.id, status);
+			}
+		});
+	},
+
+	endSession(batchId) {
+		return request(M.getImageReviewServiceUrl(`/batch/${batchId}`, {}), {
+			method: 'DELETE',
+		});
+	},
+
+	getImages(batchId) {
+		return request(M.getImageReviewServiceUrl(`/batch/${batchId}`, {}));
+	},
+
+	sanitize(rawData, batchId, userInfo, status) {
+		const images = [];
+
+		rawData.forEach((image) => {
+			if (['UNREVIEWED', 'QUESTIONABLE', 'REJECTED'].indexOf(image.currentStatus) !== -1) {
+				images.push(Ember.Object.create({
+					imageId: image.imageId,
+					fullSizeImageUrl: image.imageUrl,
+					batchId,
+					context: image.context || '#',
+					status: 'accepted',
+					history: image.imageHistory
+				}));
+			}
+		});
+
+		return ImageReviewModel.create({
+			images,
+			batchId,
+			userCanAuditReviews: userInfo,
+			status
+		});
+	},
+
+	reviewImages(images, batchId) {
+			const imageList = images.map((item) => {
+				return {
+					imageId: item.imageId,
+					imageStatus:item.status.toUpperCase()
+				};
+			});
+
+			return request(M.getImageReviewServiceUrl(`/batch/${batchId}/`), {
+				method: 'POST',
+				dataType: 'text', //this is a dirty workaround
+				data: JSON.stringify({images: imageList})
+			});
+	},
+
+	getImagesAndPermission(batchId, status) {
+		const promises = [
+			ImageReviewModel.getImages(batchId),
+			ImageReviewModel.getUserAuditReviewPermission()
+		];
+
+		return Ember.RSVP.allSettled(promises)
+		.then(([getImagesPromise, getUserAuditReviewPermissionPromise]) => {
+			return ImageReviewModel
+			.sanitize(
+					getImagesPromise.value.imageList,
+					getImagesPromise.value.batchId,
+					getUserAuditReviewPermissionPromise.value,
+					status
+			);
+		});
+	},
+
+	getUserAuditReviewPermission() {
+		return request(M.getImageReviewServiceUrl('/info', {}), {
+			method: 'GET',
+		}).then((payload) => {
+			return payload.userAllowedToAuditReviews;
+		});
+	},
+
+	createEmptyModelWithPermission(status) {
+		return ImageReviewModel.getUserAuditReviewPermission().then((userInfo) =>
+				ImageReviewModel.create({userCanAuditReviews: userInfo, status}));
+	},
+
+
+});
+
+export default ImageReviewModel;
