+++ front/main/app/routes/image-review.js
@@ -1,111 +1,2 @@
 import Ember from 'ember';
-import ImageReviewModel from '../models/image-review';
-
-const {Route, Logger} = Ember;
-
-export default Route.extend({
-	status: 'UNREVIEWED',
-
-	renderTemplate(controller, model) {
-		this.render('image-review', {controller, model});
-	},
-
-	model() {
-		return ImageReviewModel.startSession(this.get('status'));
-	},
-
-	afterModel(model) {
-		model.setImagesCount(this.get('status'));
-		this.controllerFor('application').set('isLoading', false);
-	},
-
-	actions: {
-		error(error) {
-			Logger.error('image-review route error', error);
-			let errorMessage = i18n.t('main.error-other', {ns: 'image-review'});
-
-			if (error.status === 401) {
-				errorMessage = i18n.t('main.error-no-access-permissions', {ns: 'image-review'});
-			}
-
-			this.controllerFor('application').addAlert({
-				message: errorMessage,
-				type: 'warning',
-				persistent: true
-			});
-			this.transitionTo('wiki-page', '');
-			return false;
-		},
-
-		reviewAndGetMoreImages() {
-			const model = this.modelFor('imageReview');
-			this.controllerFor('application').set('isLoading', true);
-			window.scrollTo(0, 0);
-
-			ImageReviewModel.reviewImages(model.images, model.contractId).then(() => {}, (data) => {
-				this.controllerFor('application').addAlert({
-					message: data,
-					type: 'warning',
-					persistent: false
-				});
-			}).then(this.refresh.bind(this));
-		},
-
-		getAllWithStatus(status) {
-			const model = this.modelFor('imageReview');
-			window.scrollTo(0, 0);
-
-			if (!Ember.isNone(model.contractId)) {
-				ImageReviewModel.endSession(model.contractId);
-			}
-
-			this.set('status', status);
-			this.refresh();
-		},
-
-		openMainPage() {
-			this.transitionTo('wiki-page', '');
-		},
-
-		toggleSummary() {
-			this.modelFor('image-review').setSummaryModel();
-		},
-
-		setStartDate(startDate) {
-			this.modelFor('image-review').set('startDate', startDate);
-		},
-
-		setEndDate(endDate) {
-			this.modelFor('image-review').set('endDate', endDate);
-		},
-
-		didTransition() {
-			this.controllerFor('application').set('fullPage', true);
-			if (this.controller.get('fullscreen') === 'true') {
-				this.modelFor('imageReview').set('showSubHeader', false);
-			} else {
-				this.modelFor('imageReview').set('showSubHeader', true);
-			}
-		},
-
-		changeItemModel(id, status) {
-			this.modelFor('imageReview').images.forEach((item) => {
-				if (item.imageId === id) {
-					item.set('status', status);
-				}
-			});
-		},
-
-		willTransition(transition) {
-			const isStayingOnEditor = transition.targetName.indexOf('image-review') > -1;
-
-			if (!isStayingOnEditor) {
-				transition.then(() => {
-					this.controllerFor('application').set('fullPage', false);
-				});
-			}
-
-			return true;
-		}
-	}
-});
+export default Ember.Route.extend({});
