+++ front/main/app/templates/components/image-review-summary.hbs
@@ -1,35 +1,58 @@
-<table class="image-review-modal-history">
-	<tr>
-		<td>Total Reviewed</td>
-		<td>Average Per User</td>
-		{{#each data.reviewsCountPerStatus as |countStatus|}}
-			<td>{{countStatus.status}}</td>
-		{{/each}}
-	</tr>
-	<tr>
-		<td>{{data.totalReviewed}}</td>
-		<td>{{data.avgPerUser}}</td>
-		{{#each data.reviewsCountPerStatus as |countPerStatus|}}
-			<td>{{countPerStatus.count}}</td>
-		{{/each}}
-	</tr>
-</table>
-<br>
-<table class="image-review-modal-history">
-	<tr>
-		<td>User name</td>
-		<td>Distance to avg</td>
-		{{#each data.reviewsCountPerStatus as |countStatus|}}
-			<td>{{countStatus.status}}</td>
-		{{/each}}
-	</tr>
-	{{#each data.userStatistics as |user|}}
-		<tr>
-			<td>{{user.userName}}</td>
-			<td>{{user.distanceToAvg}}</td>
-			{{#each user.reviewsCountPerStatus as |countPerStatus|}}
-				<td>{{countPerStatus.count}}</td>
-			{{/each}}
-		</tr>
-	{{/each}}
-</table>
+{{#if model.showSubHeader}}
+	<header class="sub-head">
+		<a {{action this.attrs.openImageReview}} class="sub-head--cancel">{{svg 'back-arrow' role='img' class='icon'}}</a>
+		<div class="sub-head--title">
+			{{i18n 'main.head' ns='image-review'}}
+		</div>
+	</header>
+{{/if}}
+
+{{#if model.userCanAuditReviews}}
+	<div class="image-review-summary-dialog">
+		<label class="image-review-date-input">
+			From: {{pikaday-input theme="image-review-date-picker" onSelection=(action this.attrs.setStartDate)}}📅
+		</label>
+		<label class="image-review-date-input">
+			To: {{pikaday-input theme="image-review-date-picker" onSelection=(action this.attrs.setEndDate)}}📅
+		</label>
+		<button {{action this.attrs.toggleSummary}} class="image-review-toggle-summary-button">Show statistics</button>
+		{{#if model.summaryModel}}
+			<table class="image-review-modal-history">
+				<tr>
+					<td>Total Reviewed</td>
+					<td>Average Per User</td>
+					{{#each data.reviewsCountPerStatus as |countStatus|}}
+						<td>{{countStatus.status}}</td>
+					{{/each}}
+				</tr>
+				<tr>
+					<td>{{data.totalReviewed}}</td>
+					<td>{{data.avgPerUser}}</td>
+					{{#each data.reviewsCountPerStatus as |countPerStatus|}}
+						<td>{{countPerStatus.count}}</td>
+					{{/each}}
+				</tr>
+			</table>
+			<br>
+			<table class="image-review-modal-history">
+				<tr>
+					<td>User name</td>
+					<td>Distance to avg</td>
+					{{#each data.reviewsCountPerStatus as |countStatus|}}
+						<td>{{countStatus.status}}</td>
+					{{/each}}
+				</tr>
+				{{#each data.userStatistics as |user|}}
+					<tr>
+						<td>{{user.userName}}</td>
+						<td>{{user.distanceToAvg}}</td>
+						{{#each user.reviewsCountPerStatus as |countPerStatus|}}
+							<td>{{countPerStatus.count}}</td>
+						{{/each}}
+					</tr>
+				{{/each}}
+			</table>
+
+		{{/if}}
+	</div>
+{{/if}}
