+++ front/main/app/templates/image-review.hbs
@@ -1,10 +1,1 @@
-{{image-review
-	model=model
-	reviewAndGetMoreImages=(action 'reviewAndGetMoreImages')
-	changeItemModel=(action 'changeItemModel')
-	openMainPage=(action 'openMainPage')
-	toggleSummary=(action 'toggleSummary')
-	getAllWithStatus=(action 'getAllWithStatus')
-	setStartDate=(action 'setStartDate')
-	setEndDate=(action 'setEndDate')
-}}
+{{outlet}}
