+++ front/main/app/templates/image-review/index.hbs
@@ -1,0 +1,8 @@
+{{image-review
+	model=model
+	reviewAndGetMoreImages=(action 'reviewAndGetMoreImages')
+	changeItemModel=(action 'changeItemModel')
+	openMainPage=(action 'openMainPage')
+	openSummary=(action 'openSummary')
+	getAllWithStatus=(action 'getAllWithStatus')
+}}
