+++ front/main/app/templates/image-review/summary.hbs
@@ -1,0 +1,7 @@
+{{image-review-summary
+	model=model
+	toggleSummary=(action 'toggleSummary')
+	setStartDate=(action 'setStartDate')
+	setEndDate=(action 'setEndDate')
+	openImageReview=(action 'openImageReview')
+}}
